Added Prorrata functionality to AEAT 303/config
Now prorrata is taken into account when calculating model 303 29, 31 and 41 boxes.
This value is defined on account configuration, and can be calculated through a newly added button function there.
A proportional move is created to adjust for the change in value on the boxes, when completing the 303 report.
The move is set to the account also defined in the configuration.

Task #173970