Add the option of prorating to form 303. (#11)
* Added Prorrata functionality to AEAT 303/config

Now prorrata is taken into account when calculating model 303 29, 31 and 41 boxes.
This value is defined on account configuration, and can be calculated through a newly added button function there.
A proportional move is created to adjust for the change in value on the boxes, when completing the 303 report.
The move is set to the account also defined in the configuration.

Task #173970

* Additional prorrata_account check.

Required if prorrata_percent is set -> no prorrata 303 report can be processed without account.

Task #173970

* Corrected mistake in prorrata update chart

Task #173970

* Temporary commit
Started development special prorrata for the current year, and compare if it matches with the applied prorrata applied during the year.
Changed the way prorrata is calculated, now it takes into account the periods of the specified fiscal year.

Task #173970

* Development as specified by task 0668912
Changed the way prorrrata is calculated; now it uses the periods of the defined fiscal year under aeat303 account configuration.
Last period mode allowed, that calculates prorrata for the whole and current fiscal year, to be used when doing last period aeat 303 report.
303 report also changes, as now takes into account the prorrata applied set on account configuration and the difference with the last period prorrata, to fill field 44.

Task #173970

* Expanded tests and translation

Now tests also cover the case where is the last period of fiscal year, and prorrata has to be recalculated with the yearly fiscal accountability.

Task #173970

* - Improve the way the prorrata percentage is calculated.
- Set the prorrata percentage when the last 303 is process and unset when
is canceled.
- Update tests.

Task: #173970

---------

Co-authored-by: Kevin Ortega Garcia <kevin.ortega@nan-tic.com>
Co-authored-by: Bernat Brunet <bernat@nan-tic.com>